add a new script to support polymer v1
Since polymer doesn't support es6 yet and the bundle mechanism
tool (that is, vulcanise) throw an error when using an async
function or method.

A class with the same name was created with another script name.
This can be use use exactly as before, the only difference is
that this new class uses promise directly instead of async.